LayoutTests/imported/w3c/web-platform-tests/2dcontext/wide-gamut-canvas/imageData-colorManagedBehavior-expected.txt

CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'srgbImageDataU8.dataUnion[i*4+j] = srgbPixels[i][j]')

Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'srgbImageDataU8.dataUnion[i*4+j] = srgbPixels[i][j]')
